requirements.txt
asgiref==3.8.1
asttokens==3.0.0
cascade==0.1
certifi==2025.4.26
charset-normalizer==3.4.2
coreapi==2.3.3
coreschema==0.0.4
coverage==7.8.0
decorator==5.2.1
Django==4.2.10
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
executing==2.2.0
forex-python==1.8
idna==3.10
imaplib2==3.6
inflection==0.5.1
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itypes==1.2.0
jedi==0.19.2
Jinja2==3.1.6
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
progressbar==2.5
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.9.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
setuptools==80.7.1
simplejson==3.20.1
sqlparse==0.5.3
stack-data==0.6.3
stripe==12.1.0
traitlets==5.14.3
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
wcwidth==0.2.13
wheel==0.45.1
